feat: heartbeat-generated teletext samples
The TextChunker that produces segments is triggered by
TextSamples, but there may be long times with no text
cues and only heartbeat stuffing data with timestamps.

This PR uses the heartbeat timestamps to generate zero-duration
TextSamples, so that the TextChunker can progress with segment
generation.

The TS teletext parser is updated to generate zero-duration
TextSamples once 0.5s has passed since last TextSample end.
It will also warn if there is gap between consecutive timestamps
longer than 5s.

The TextChunker is updated to drop such heartbeat TextSamples,
but progress with segment generation with based on start time.
This avoids that the text sample generation stalls when
no new teletext cues are received.

There has also been an optimization to update the start time
of a teletext cue when the first text row is received.

Finally, the TextPadder is disabled for now, as it does
not work well for live timestamps not starting at 0.